+++ src/leiningen/new/auth_base.clj
@@ -6,10 +6,10 @@
     [assets
      (-> options
          (assoc :auth true)
-         (append-options :dependencies [['buddy/buddy-auth "3.0.1"]
-                                        ['buddy/buddy-sign "3.4.1"]
-                                        ['buddy/buddy-core "1.10.1"]
-                                        ['buddy/buddy-hashers "1.8.1"]])
+         (append-options :dependencies [['buddy/buddy-auth "3.0.323"]
+                                        ['buddy/buddy-sign "3.4.333"]
+                                        ['buddy/buddy-core "1.10.413"]
+                                        ['buddy/buddy-hashers "1.8.158"]])
          (append-formatted :auth-middleware-required
                            [['buddy.auth.middleware :refer ['wrap-authentication 'wrap-authorization]]
                             ['buddy.auth.accessrules :refer ['restrict]]