+++ src/leiningen/new/shadow_cljs.clj
@@ -1,10 +1,10 @@
 (ns leiningen.new.shadow-cljs
   (:require [leiningen.new.common :refer :all]))
 
-(def shadow-version "2.16.7")
+(def shadow-version "2.17.0")
 
 (def shadow-cljs-dependencies [['thheller/shadow-cljs shadow-version :scope "provided"]
-                               ['org.clojure/core.async "1.5.644"]])
+                               ['org.clojure/core.async "1.5.648"]])
 
 (def shadow-assets
   [["package.json" "shadow/package.json"]